Steganography/requirements.txt
numpy==1.20.1
Pillow==8.1.1
opencv-python~=4.6.0.66
docopt~=0.6.2